Adding partial course listing web scraper
(*) course_search_keys generates a unique session id and browser
refresh by scraping with a GET request and picking out the
__BrowserRefresh field value and ASP.NET_SessionId set-cookie response
header.
(*) courses-to-jsonFile uses this to simulate a POST request (currently
one that only requests the first 20 listed courses — can’t request more
than the first page of courses right now) handled by course-processor
(*) course-processor makes the http request and parses the HTML,
returning an array to a courses-to-jsonFile handler, which serializes
the data in a JSON format and saves it to course-data.json.

Still need to figure out how to have the scraper be able to navigate
the pages of later results…