Full course scraper complete!
Code should be cleaned up as usual and parameters shouldn’t be
hardcoded, but all 62 pages of course data for the Spring 2015 semester
have now been scraped using the course processor.  A small fix can make
it scrape to determine the number of pages.  It goes through the pages
by repeatedly getting the new __BrowserRefresh value while maintaining
the session key (consistent session key seems to be required to do
this).