Added EJS view for Course Search
You can now see the list of courses with some mistakes because there
are some problems with the parsing:

(*) Only first professor name is saved.
(*) Nursing courses with, “No schedule is currently available” are not
tagged consistently (probably because the program only adds li tags
when data is present for those fields) so the data is being processed
incorrectly for some courses.