Removing public/data/articles.json from .gitignore
In the future, we should probably have the processor script run automatically to get new articles and cache the results, but for now, the Articles page is rendered from the articles.json file.  Although we can change this, I don't think we should until the processor script is capable of detecting potential formatting changes to the Molloy Life News Page that would break it.